latency flag added;

+++ src/entrypoint.cc
@@ -281,7 +281,7 @@ int user_entrypoint(std::shared_ptr<utility::parse::Config> config, boost::progr
       }
 
       std::map<std::string, std::vector<std::string>> results;
-      for (int i = 0; i < str_addrs.size(); i++) {
+      for (int i = 0; i < (int)str_addrs.size(); i++) {
         google::protobuf::Map<string, string> r = db_addrs[i]->get_db();
         results[str_addrs[i]];
         for (std::string key : keys) {
